requirements.txt
aiogram>=3,<4
supabase>=2,<3
python-dotenv>=1,<2

